+++ build_examples.sh
@@ -2,5 +2,5 @@
 
 set -eux
 
-cargo build --release --target=thumbv7em-none-eabi --examples
-cargo build --release --target=riscv32imc-unknown-none-elf --examples # Important for tests: This target does not support atomics
+PLATFORM=nrf52 cargo build --release --target=thumbv7em-none-eabi --examples
+PLATFORM=riscv32 cargo build --release --target=riscv32imc-unknown-none-elf --examples # Important for testing: This target does not support atomics