+++ run_all_checks.sh
@@ -2,6 +2,8 @@
 
 set -eux
 
+export PLATFORM=nrf52 # The specific platform doesn't matter for tests
+
 cargo fmt --all -- --check
 cargo test --workspace
 cargo clippy --workspace --all-targets